+++ changelog.md
@@ -52,6 +52,8 @@ becomes an alias for `addr`.
 
 - Deprecated `selfExe` for Nimscript.
 
+- 'lock levels' is deprecated, now a noop.
+
 
 ## Language changes